arreglo de modularizacion
