+++ python/SEGMENTACION DE CLIENTES/JSONprueba.py
@@ -1,4 +1,4 @@
-
+from moduloCliente import Cliente_black,Cliente_clasico,Cliente_gold
 from copy import copy
 import json
 
@@ -43,7 +43,30 @@ class Json(object):
     def obtenerDireccion(self,dato):
         with open(self.archivo,"r") as file:
             datos = json.load(file)
-            return datos["direccion"][dato]       
+            return datos["direccion"][dato]      
+        
+        
+def iniciarPrograma(archivo):
+    archivo = archivo.upper()
+    
+    if archivo == "CLASSIC":
+        print(archivo)
+        archivo = eventos_classic
+        cargado = Json(archivo)
+        return Cliente_clasico(cargado,nombre=cargado.nombre,apellido=cargado.apellido,numero=cargado.apellido,dni=cargado.dni,tipo=cargado.tipo)
+    elif archivo == "GOLD":
+        print(archivo)
+        archivo = eventos_gold
+        cargado = Json(archivo)
+        return Cliente_gold(cargado,nombre=cargado.nombre,apellido=cargado.apellido,numero=cargado.apellido,dni=cargado.dni,tipo=cargado.tipo)
+    elif archivo == "BLACK":
+        print(archivo)
+        archivo = eventos_black
+        cargado = Json(archivo)
+        return Cliente_black(cargado,nombre=cargado.nombre,apellido=cargado.apellido,numero=cargado.apellido,dni=cargado.dni,tipo=cargado.tipo)
+    else:
+        print("archivo no reconocido")
+
