+++ python/SEGMENTACION DE CLIENTES/PruebaFinal.py
@@ -1,4 +1,4 @@
-from JSONprueba import eventos_black,eventos_classic,eventos_gold
+from JSONprueba import eventos_black,eventos_classic,eventos_gold, iniciarPrograma
 from moduloCliente import Cliente_black,Cliente_clasico,Cliente_gold
 from print_HTML import plantillaHTML
 from JSONprueba import Json
@@ -8,29 +8,8 @@ from JSONprueba import Json
 #-------------------------Ingrese el archivo que se desea escanear (eventos_classic,eventos_gold,eventos_black)---------------------------------------------------------#
 archivo = input("Ingrese el archivo que desea leer: ")
 
-def iniciarPrograma(archivo):
-    archivo = archivo.upper()
-    
-    if archivo == "CLASSIC":
-        print(archivo)
-        archivo = eventos_classic
-        cargado = Json(archivo)
-        return Cliente_clasico(cargado,nombre=cargado.nombre,apellido=cargado.apellido,numero=cargado.apellido,dni=cargado.dni,tipo=cargado.tipo)
-    elif archivo == "GOLD":
-        print(archivo)
-        archivo = eventos_gold
-        cargado = Json(archivo)
-        return Cliente_gold(cargado,nombre=cargado.nombre,apellido=cargado.apellido,numero=cargado.apellido,dni=cargado.dni,tipo=cargado.tipo)
-    elif archivo == "BLACK":
-        print(archivo)
-        archivo = eventos_black
-        cargado = Json(archivo)
-        return Cliente_black(cargado,nombre=cargado.nombre,apellido=cargado.apellido,numero=cargado.apellido,dni=cargado.dni,tipo=cargado.tipo)
-    else:
-        print("archivo no reconocido")
 
 
- 
 x = iniciarPrograma(archivo)
 plantillaHTML(x)
 #--------------------------El archivo HTML se llama: indexpy.html (Caprpeta "SEGMENTACION CLIENTES"-------------------------------------------------------------------------------------------------------#
